+++ src/EmbedReactNative.ts
@@ -107,6 +107,7 @@ export type Gr4vyConfig = {
   requireSecurityCode?: boolean
   shippingDetailsId?: string
   merchantAccountId?: string
+  connectionOptions?: Record<string, unknown>
   debugMode?: boolean
 }