+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# v1.5.1 (Wed Aug 14 2024)
+
+#### 🐛 Bug Fix
+
+- chore: update docs with connection options [#39](https://github.com/gr4vy/gr4vy-react-native/pull/39) ([@luca-gr4vy](https://github.com/luca-gr4vy))
+
+#### Authors: 1
+
+- Luca Allievi ([@luca-gr4vy](https://github.com/luca-gr4vy))
+
+---
+
+# v1.5.0 (Tue Aug 13 2024)
+
+#### 🚀 Enhancement
+
+- task: add support for `connectionOptions` [#37](https://github.com/gr4vy/gr4vy-react-native/pull/37) ([@luca-gr4vy](https://github.com/luca-gr4vy))
+
+#### Authors: 1
+
+- Luca Allievi ([@luca-gr4vy](https://github.com/luca-gr4vy))
+
+---
+
 # v1.4.0 (Tue Jul 02 2024)
 
 #### 🚀 Enhancement